#50 Make TextGraphWriter private and test directly TextGraphFormatter
